refactor: updated middleware to allow universal liking and commenting, and added protections against unauthorized operations in Prisma queries

+++ services/CommentService.ts
@@ -1,8 +1,17 @@
+import { getServerSession } from "next-auth";
+
+import { authOptions } from "../app/api/auth/[...nextauth]/options";
 import { CommentCRUD } from "../models/Comment";
 import prisma from "../utils/prisma";
 
 export const createComment = async (commentData: CommentCRUD) => {
+  const userSession = await getServerSession(authOptions);
+
   const { userId, postId, ...rest } = commentData;
+
+  if (userSession!.user.id !== userId) {
+    throw new Error("Forbidden: Operation on behalf of another user");
+  }
 
   const comment = await prisma.comment.create({
     data: {
@@ -35,6 +44,16 @@ export const getCommentsByPostId = async (postId: number) => {
 };
 
 export const deleteComment = async (commentId: number) => {
+  const userSession = await getServerSession(authOptions);
+
+  const comment = await prisma.comment.findUnique({
+    where: { id: commentId },
+  });
+
+  if (userSession!.user.id !== comment!.userId) {
+    throw new Error("Forbidden: Operation on behalf of another user");
+  }
+
   const deletedComment = await prisma.comment.delete({
     where: { id: commentId },
   });
